bug fixes and performance improvements.
Just a grab-bag of random fixes I had sitting on my computer uncommitted
that I'm too lazy to break up into coherent commits.

Update our go.mod to use local versions of the other auth modules, like
some kind of sociopath.

Default Grants created from Google IDs to unused, so we can mark them as
used after they're exchanged for a token.

Catch when we're trying to redirect without a redirect URL and handle
it.

Actually validate client credentials.

Actually retrieve client redirect URIs.

Return the grant that's created from our createGrant helper, so we can
get the version with defaults filled in. Also, catch when the grant
source has already been used, and return an invalid grant error.

Add a placeholder so we remember to implement our email flow.

Build a running log inside our access token handler so it can accrete
information as we get further into the processing of a request.

Fill out error handling when trying to exchange a grant.

Make refresh token expiration configured through our Service type.

Default refresh token-based grants to unused, so they can be marked as
used when exchanged for a grant.

Make our logger-injection middleware not dependent on the Service type
anymore, as it's going to be relying on an upstream middleware to inject
the logger, and it's only going to inject the endpoint and request
method, along with request vars, now.