Add authorization endpoint, email flow.
Add a new authorization endpoint for generating grants that use an OOB
or redirect flow.

Add an email grant handler to generate a Grant, then email it to users,
to be clicked as the authentication step.

I'm not super happy with any of these abstractions, really, but this
probably works for the moment.