Update email grants to use randomly generated code.
Randomly generate the code that will be sent to the user's email address
to verify account ownership when using the email grant method. We'll
retrieve the grant by unique code, and pull the email address out of it,
instead of parsing the email out of the code, which is more brittle and
less user-friendly.