Fix scopes.
Add more thorough testing by checking what's persisted on the grant
that's stored in the grants.Storer.

This revealed that we weren't persisting the scopes on the grant
correctly because the scopes were getting parsed as [""] instead of [],
meaning we weren't populating default scope values. This also fixes that
bug.