Finish testing grant creation from email, fix bugs.
Add test cases checking that scopes are handled correctly when creating
grants using emails, and fix the bug exposed by them, namely that we
weren't validating that users had access to the scopes they were
requesting; we only checked clients. This involved updating the scopes
module, but also involved refactoring our granters and grantCreators to
be able to surface the profile ID a grant is for, which was annoying.

In theory, in the future, it may be desirable to filter scopes based on
the account ID (e.g., you can request admin scopes using your work
email, but they won't be handed out if you request them with your
personal email, or you can request admin scopes using your email but if
you're logging in with WebAuthn, you don't get admin scopes, or
whatever) but the big issue here is that refresh tokens aren't currently
tracking the account that created them, only the grant. So to get to the
account that started the chain, you need to unwrap the token into the
grant that created it, which has that information. Unfortunately, when
creating the Grant that gets exchanged for the refresh token, we're not
populating the account ID on the grant because we don't have easy access
to it. That's probably not good, and we should address that.

🎵 Now Playing: Impossible Year
🎵 Artist: Panic! at the Disco
🎵 Album: Death of a Bachelor