Start testing our token handler.
Add tests for some basic HTTP level tests of the token handler. Fix the
content-type problem that has plagued us before, and make sure we're
returning the appropriate error when clients validation fails; before we
were always treating it as an invalid request, but it may have been a
server error.

🎵 Now Playing: Yuri on ICE
🎵 Artist: 梅林太郎
🎵 Album: Oh! スケトラ!!! ユーリ!!! on ICE オリジナル・スケートソングCOLLECTION