Add some more tests.
Add a few more tests for our oauth2 handlers. Ran into a bit of a pickle
in that we need to know what tokens are going to be issued before
they're actually issued, but we can't know that, as that involves some
random data.

May be time for some mocks to be introduced? Unclear.